refactor: BridgeGameResult를 vo 패키지로 이동
