+++ src/main/java/bridge/view/OutputView.java
@@ -1,16 +1,13 @@
 package bridge.view;
 
-import bridge.domain.BridgeGameResult;
-import bridge.domain.enumeration.MovingType;
+import bridge.domain.vo.BridgeGameResult;
 
 import java.util.List;
-import java.util.Objects;
 
 import static bridge.domain.BridgeGame.RETRY_COUNT;
-import static bridge.domain.BridgeGameResult.*;
+import static bridge.domain.vo.BridgeGameResult.*;
 import static bridge.domain.enumeration.Format.*;
 import static bridge.domain.enumeration.Message.*;
-import static bridge.domain.enumeration.MovingType.*;
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
